Consolidate connect code into single connect.js;
